+++ flowcelltool/flowcells/views.py
@@ -707,6 +707,37 @@ class FlowCellExtractLibrariesView(
             return redirect(reverse(
                 'flowcell_view', kwargs={'uuid': flow_cell.uuid}))
 
+    def render_done(self, form, **kwargs):
+        """We need to reproduce this here for error handling in the ``done()`` method.
+        See the following for more information:
+
+        - https://github.com/django/django-formtools/issues/61
+        """
+        from collections import OrderedDict
+        final_forms = OrderedDict()
+        # walk through the form list and try to validate the data again.
+        for form_key in self.get_form_list():
+            form_obj = self.get_form(
+                step=form_key,
+                data=self.storage.get_step_data(form_key),
+                files=self.storage.get_step_files(form_key)
+            )
+            if not form_obj.is_valid():
+                return self.render_revalidation_failure(form_key, form_obj, **kwargs)
+            final_forms[form_key] = form_obj
+
+        # render the done view and reset the wizard before returning the
+        # response. This is needed to prevent from rendering done with the
+        # same data twice.
+        from django.core.exceptions import ValidationError
+        try:
+            done_response = self.done(final_forms.values(), form_dict=final_forms, **kwargs)
+            self.storage.reset()
+            return done_response
+        except ValidationError as e:
+            form.add_error(None, e)
+            return self.render(form)
+
     def _build_libraries(self):
         """Build library objects from result of paste_tsv and pick_columns
         steps