+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 HEAD
 ----
 
+- Fixing issue with validation errors late in the XLSX import (#141, #142).
 - Adding ``manual_label`` to ``FlowCell`` (#147).
 - Adding run number to flowcell list.
 - Fixing flow cell order.